+++ trace/retry.go
@@ -15,7 +15,11 @@ type (
 		OnRetry func(RetryLoopStartInfo) func(RetryLoopIntermediateInfo) func(RetryLoopDoneInfo)
 	}
 	RetryLoopStartInfo struct {
-		Context    context.Context
+		// Context make available context in trace callback function.
+		// Pointer to context provide replacement of context in trace callback function.
+		// Warning: concurrent access to pointer on client side must be excluded.
+		// Safe replacement of context are provided only inside callback function
+		Context    *context.Context
 		ID         string
 		Idempotent bool
 	}